requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
black==23.11.0
celery==5.3.6
certifi==2025.8.3
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.3.4
cryptography==45.0.6
daphne==4.0.0
Django==5.2.5
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.6.0
flake8==6.1.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
isort==5.12.0
kombu==5.5.4
loguru==0.7.3
Markdown==3.8.2
mccabe==0.7.0
msgpack==1.1.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.1.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.0.8
sentry-sdk==1.38.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.6.0
zope.interface==7.2
